Move the test into isolate_test
